ride_sharing_backend/requirements.txt
django 
djangorestframework 
djangorestframework-simplejwt 
django-cors-headers 
Pillow 
django-phonenumber-field 
phonenumbers
